2023 Day 24

+++ src/main/scala/year_2023/day_024/Hailstone.scala
@@ -1,5 +1,5 @@
 package year_2023.day_024
 
 
-case class Coordinate(x:Double, y:Double, z:Double)
+case class Coordinate(x:Long, y:Long, z:Long)
 case class Hailstone(initial:Coordinate, speed:Coordinate)
